% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{AIC.sgdgmf}
\alias{AIC.sgdgmf}
\title{Compute the AIC of a GMF model}
\usage{
\method{AIC}{sgdgmf}(object)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\description{
Compute the Akaike information criterion (AIC) of an estimated GMF object
}
