% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{BIC.sgdgmf}
\alias{BIC.sgdgmf}
\title{Compute the BIC of a GMF model}
\usage{
\method{BIC}{sgdgmf}(object)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\description{
Compute the Bayesian information criterion (BIC) of an estimated GMF object
}
