% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGMF-class.R
\name{biplot.initgmf}
\alias{biplot.initgmf}
\title{Biplot of an initialized GMF model}
\usage{
\method{biplot}{initgmf}(object, choices = 1:2, normalize = FALSE, labels = NULL, palette = NULL)
}
\arguments{
\item{object}{an object of class \code{initgmf}}

\item{choices}{a length 2 vector specifying the components to plot}

\item{normalize}{if \code{TRUE}, orthogonalizes the scores using SVD}

\item{labels}{a vector of labels which should be plotted}

\item{palette}{the color-palette which should be used}
}
\description{
Plot the observations on a two-dimensional projection determined by the
estimated score matrix
}
\examples{
library(sgdGMF)

# Generate data from a Poisson model
data = sim.gmf.data(n = 100, m = 20, ncomp = 5, family = poisson())

# Fit a GMF model
init = init.gmf.param(data$Y, ncomp = 3, family = poisson())

# Get the biplot of a GMF model
biplot(init) # 1st vs 2nd principal components
biplot(init, choices = 2:3) #2nd vs 3rd principal components

}
\seealso{
\code{\link{biplot.sgdgmf}}.
}
