% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{biplot.sgdgmf}
\alias{biplot.sgdgmf}
\title{Biplot of a GMF model}
\usage{
\method{biplot}{sgdgmf}(
  object,
  choices = 1:2,
  normalize = FALSE,
  labels = NULL,
  palette = c("viridis", "inferno")
)
}
\arguments{
\item{object}{...}

\item{choices}{...}

\item{normalize}{...}

\item{labels}{...}

\item{palette}{...}
}
\description{
A short description...
}
