% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGMF-class.R
\name{coefficients.initgmf}
\alias{coefficients.initgmf}
\alias{coef.initgmf}
\title{Extract the coefficient of an initialized GMF model}
\usage{
\method{coefficients}{initgmf}(object, type = c("all", "colreg", "rowreg", "scores", "loadings"))

\method{coef}{initgmf}(object, type = c("all", "colreg", "rowreg", "scores", "loadings"))
}
\arguments{
\item{object}{an object of class \code{initgmf}}

\item{type}{the type of coefficients which should be returned}
}
\description{
Return the initialized coefficients of a GMF model, i.e., the row- and column-specific
regression effects, the latent scores and loadings.
}
\seealso{
\code{\link{coefficients.sgdgmf}} and \code{\link{coef.sgdgmf}}.
}
