% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coefficients}
\alias{coefficients}
\alias{coefficients.sgdgmf}
\title{Extract the coefficient of a GMF model}
\usage{
coefficients(object)

\method{coefficients}{sgdgmf}(
  object,
  type = c("all", "col-reg", "row-reg", "scores", "loadings")
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\value{
\code{NULL}
}
\description{
...
}
