% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct.uv}
\alias{correct.uv}
\title{Rotate the matrices U and V}
\usage{
correct.uv(U, V)
}
\description{
Rotate U and V in such a way that the transformed matrices
are such that U is orthogonal and V is lower triangular
}
\keyword{internal}
