% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp.airwls.glmstep}
\alias{cpp.airwls.glmstep}
\title{Compute one Fisher scoring step for GLMs}
\usage{
cpp.airwls.glmstep(beta, y, X, familyname, linkname, offset, weights, penalty)
}
\arguments{
\item{beta}{current value of the regression coefficients to be updated}

\item{y}{response vector}

\item{X}{design matrix}

\item{familyname}{model family name}

\item{linkname}{link function name}

\item{offset}{vector of constants to be added to the linear predictor}

\item{weights}{vector of constants non-negative weights}

\item{penalty}{penalty parameter of a ridge-type penalty}
}
\description{
Internal function to compute one Fisher scoring step for GLMs.
It constitutes the building block of the AIRWLS algorithm for the
estimation of GMF models.
}
\keyword{internal}
