% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{custom.init}
\alias{custom.init}
\title{SVD initialization}
\usage{
custom.init(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  values = list(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
