% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{deviance.sgdgmf}
\alias{deviance.sgdgmf}
\title{Compute the deviance of a GMF model}
\usage{
\method{deviance}{sgdgmf}(object, normalize = FALSE)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\description{
Compute the deviance of an estimated GMF object
}
