% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengap.R
\name{eigengap.act}
\alias{eigengap.act}
\title{Rank selection via adjust correlation thresholding}
\usage{
eigengap.act(covmat, nobs, maxcomp = NULL)
}
\arguments{
\item{covmat}{matrix to be decomposed}

\item{maxcomp}{maximum number of eigenvalues to compute}

\item{maxiter}{maximum number of iterations}
}
\description{
Select the number of significant principal components of a matrix via adjust
correlation threshold (ACT)
}
\references{
Fan, J., Guo, j. and Zheng, S. (2020).
\emph{Estimating number of factors by adjusted eigenvalues thresholding.}
Journal of the American Statistical Association, 117(538): 852--861
}
\keyword{internal}
