% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengap.R
\name{eigengap.evr}
\alias{eigengap.evr}
\title{Rank selection via eigenvalue ratio maximization}
\usage{
eigengap.evr(covmat, maxcomp = 50, thr = 0.95)
}
\arguments{
\item{covmat}{matrix to be decomposed}

\item{maxcomp}{maximum number of eigenvalues to compute}
}
\description{
Select the number of significant principal components of a matrix via the
eigenvalue ratio (EVR) maximization method
}
\references{
Ahn, S.C., Horenstein, A.R. (2013).
\emph{Eigenvalue ratio test for the number of factors.}
Econometrica, 81, 1203-1227
}
\keyword{internal}
