% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengap.R
\name{eigengap.oht}
\alias{eigengap.oht}
\title{Rank selection via optimal hard thresholding}
\usage{
eigengap.oht(covmat, nobs, maxcomp = NULL)
}
\arguments{
\item{covmat}{matrix to be decomposed}

\item{nobs}{number of observations used to compute the covariance matrix}

\item{maxcomp}{maximum number of eigenvalues to compute}
}
\description{
Select the number of significant principal components of a matrix via optimal
hard thresholding (OHT)
}
\references{
Gavish, M., Donoho, D.L. (2014)
\emph{The optimal hard thresholding for singular values is 4/sqrt(3).}
IEEE Transactions on Information Theory, 60(8): 5040--5053
}
\keyword{internal}
