% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengap.R
\name{eigengap.onatski}
\alias{eigengap.onatski}
\title{Rank selection via the Onatski method}
\usage{
eigengap.onatski(Y, maxcomp = 50, maxiter = 100)
}
\arguments{
\item{Y}{matrix to be decomposed}

\item{maxcomp}{maximum number of eigenvalues to compute}

\item{maxiter}{maximum number of iterations}
}
\description{
Select the number of significant principal components of a matrix via the
Onatski method
}
\references{
Onatski, A. (2010).
\emph{Determining the number of factors from empirical distribution of eigenvalues.}
Review of Economics and Statistics, 92(4): 1004-1016
}
\keyword{internal}
