% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{eigenval}
\alias{eigenval}
\title{Object eigenvalues}
\usage{
eigenval(object, ...)
}
\arguments{
\item{object}{an object to extract the eivenalues from}

\item{...}{additional arguments passed to or from other methods}
}
\description{
Generic function to compute and return the eigenvalues of a matrix
}
