% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{fitted.sgdgmf}
\alias{fitted.sgdgmf}
\title{Extract the fitted values of a GMF models}
\usage{
\method{fitted}{sgdgmf}(object, type = c("link", "response", "terms"))
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{type}{the type of fitted values which should be returned}
}
\description{
Computes the predictions from a fitted generalized matrix factorization model (GMF).
}
