% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_airwls.R
\name{glm.basic}
\alias{glm.basic}
\title{GLM basic update}
\usage{
glm.basic(
  X,
  Y,
  beta = NULL,
  family = gaussian(),
  tol = 1e-05,
  offset = 0,
  stepsize = 1,
  penalized = 0,
  steps = 1
)
}
\description{
...
}
\keyword{internal}
