% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{glm.svd.init}
\alias{glm.svd.init}
\title{GLM-SVD initialization}
\usage{
glm.svd.init(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
