% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_airwls.R
\name{gmf.airwls}
\alias{gmf.airwls}
\title{Factorize a matrix of non-Gaussian observations}
\usage{
gmf.airwls(
  Y,
  X = NULL,
  Z = NULL,
  family = poisson(),
  ncomp = 2,
  intercept = c(FALSE, FALSE),
  dispersion = FALSE,
  penalty = list(),
  init = list(),
  control = list()
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column fixed effects (\eqn{q \times m})}

\item{family}{a family as in the \code{\link{glm}} interface}

\item{ncomp}{number of random effects to estimate (default 2)}

\item{intercept}{...}

\item{dispersion}{...}

\item{penalty}{...}

\item{init}{...}

\item{control}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
