% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_block_sgd.R
\name{gmf.block.sgd}
\alias{gmf.block.sgd}
\title{Factorize a matrix of non-Gaussian observations using averaged SGD}
\usage{
gmf.block.sgd(
  Y,
  X = NULL,
  Z = NULL,
  family = gaussian(),
  ncomp = 2,
  intercept = c(FALSE, FALSE),
  dispersion = FALSE,
  penalty_u = 1,
  penalty_v = 0,
  penalty_b = 0,
  penalty = list(),
  init = list(),
  control = list()
)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{Z}{...}

\item{family}{...}

\item{ncomp}{...}

\item{intercept}{...}

\item{dispersion}{...}

\item{penalty}{...}

\item{init}{...}

\item{control}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
