% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{gmf.init.custom}
\alias{gmf.init.custom}
\title{SVD initialization}
\usage{
gmf.init.custom(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  values = list(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
