% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{gmf.init.glm}
\alias{gmf.init.glm}
\title{GLM-SVD initialization}
\usage{
gmf.init.glm(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
