% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{gmf.init}
\alias{gmf.init}
\title{Initialization of the generalized matrix factorization model}
\usage{
gmf.init(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = gaussian(),
  method = "svd",
  niter = 0,
  values = list(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
