% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{gmf.init.svd}
\alias{gmf.init.svd}
\title{OLS-SVD initialization}
\usage{
gmf.init.svd(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  niter = 0,
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
