% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_memo_sgd.R
\name{gmf.memo.sgd}
\alias{gmf.memo.sgd}
\title{Factorize a matrix of non-Gaussian observations using memoized SGD}
\usage{
gmf.memo.sgd(
  Y,
  X = NULL,
  Z = NULL,
  family = gaussian(),
  ncomp = 2,
  intercept = FALSE,
  dispersion = FALSE,
  penalty = list(),
  init = list(),
  control = list()
)
}
\arguments{
\item{Y}{...}

\item{X}{...}

\item{Z}{...}

\item{family}{...}

\item{ncomp}{...}

\item{intercept}{...}

\item{dispersion}{...}

\item{penalty}{...}

\item{init}{...}

\item{control}{...}
}
\value{
...
}
\description{
Fit a penalized generalized matrix factorization model using a coordinate
averaged stochastic gradient descent with memoized updates
}
\details{
...
}
\examples{
...

}
\references{
...
}
