% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{gmf.simulate}
\alias{gmf.simulate}
\title{Simulate a matrix of non-Gaussian observations}
\usage{
gmf.simulate(
  n = 100,
  m = 20,
  ncomp = 2,
  ngroup = 3,
  nx = NULL,
  nz = NULL,
  intercept_row = FALSE,
  intercept_col = FALSE,
  weight = 0.3,
  offset = NULL,
  family = poisson()
)
}
\arguments{
\item{n}{...}

\item{m}{...}

\item{ncomp}{...}

\item{ngroup}{...}

\item{nx}{...}

\item{nz}{...}

\item{intercept_row}{...}

\item{intercept_col}{...}

\item{weight}{...}

\item{offset}{...}

\item{family}{...}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
