% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{heatmap.sgdgmf}
\alias{heatmap.sgdgmf}
\title{Heatmap of a GMF model}
\usage{
heatmap.sgdgmf(
  object,
  type = c("data", "response", "link", "scores", "loadings", "deviance", "pearson",
    "working"),
  resid = FALSE,
  symmetric = FALSE,
  transpose = FALSE,
  limits = NULL,
  palette = NULL
)
}
\arguments{
\item{object}{...}

\item{type}{...}

\item{resid}{...}

\item{symmetric}{...}

\item{transpose}{...}

\item{limits}{...}

\item{palette}{...}
}
\description{
A short description...
}
