% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.gmf.param}
\alias{init.gmf.param}
\alias{init.param.ols}
\alias{init.param.glm}
\alias{init.param.random}
\alias{init.param.custom}
\title{Initialize the parameters of a generalized matrix factorization model}
\usage{
init.gmf.param(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  method = c("svd", "ols", "glm", "random", "values"),
  type = c("deviance", "pearson", "working", "link"),
  niter = 0,
  values = list(),
  verbose = FALSE,
  parallel = FALSE,
  nthreads = 1,
  savedata = TRUE
)

init.param.ols(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  type = c("deviance", "pearson", "working", "link"),
  verbose = FALSE
)

init.param.glm(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  type = c("deviance", "pearson", "working", "link"),
  verbose = FALSE,
  parallel = FALSE,
  nthreads = 1
)

init.param.random(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  sigma = 1
)

init.param.custom(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  values = list(),
  verbose = FALSE
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row-specific fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column-specific fixed effects (\eqn{q \times m})}

\item{ncomp}{rank of the latent matrix factorization}

\item{family}{a family as in the \code{\link{glm}} interface}

\item{method}{optimization method to be used for the initial fit}

\item{type}{type of residuals to be used for initializing \code{U} via incomplete SVD decomposition}

\item{niter}{number of iterations to refine the initial estimate (only if \code{method="ols"} or \code{"svd"})}

\item{values}{a list of custom initial values for \code{B}, \code{A}, \code{U} and \code{V}}

\item{verbose}{print the status of the initialization process}

\item{parallel}{if \code{TRUE}, allows for parallel computing using the \code{foreach} package}

\item{nthreads}{number of cores to be used in parallel (only if \code{parallel=TRUE})}

\item{savedata}{if \code{TRUE}, stores a copy of the input data}
}
\value{
A list containing the initial values of \code{B}, \code{A}, \code{U}, \code{V} and \code{phi}.
}
\description{
Provide four initialization methods to set the initial values of the parameters of
a generalized matrix factorization (GMF) model identified by a \code{\link{glm}} family
and a linear predictor of the form \eqn{g(\mu) = \eta = X B^\top + A Z^\top + U V^\top},
with bijective link function \eqn{g(\cdot)}.
See \code{\link{sgdgmf}} for more details on the model specification.
}
\details{
If \code{method = "svd"}, the initialization is performed fitting a sequence of linear
regressions followed by a residual SVD decomposition.
To account for non-Gaussian distribution of the data, regression and
decomposition are applied on the transformed response matrix \eqn{Y_h = (g \circ h)(Y)},
where \eqn{h(\cdot)} is a function which prevent \eqn{Y_h} to take infinite values.
For instance, in the Binomial case \eqn{h(y) = 2 (1-\epsilon) y + \epsilon},
while in the Poisson case \eqn{h(y) = y + \epsilon}, where \eqn{\epsilon} is a small
positive constant, typically \code{0.1} or \code{0.01}.

If \code{method = "glm"}, the initialization is performed by fitting a sequence of
generalized linear models followed by a residual SVD decomposition.
In particular, we use independent GLM fit with \eqn{y_j \sim X \beta_j} to set \eqn{\beta_j}.
Similarly, we fit the model \eqn{y_i \sim Z \alpha_i + o_i} with offset \eqn{o_i = B x_i}
to set \eqn{\alpha_j}. Then, \eqn{U} and \eqn{V} are obtained via SVD on the final
working residuals.

If  \code{method = "random"}, the initialization is performed using independent Gaussian
random values for all the parameters in the model

If \code{method = "values"}, the initialization is performed using user-specified
values provided as an input.
}
\examples{
...

}
\keyword{internal}
