% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init.param.custom}
\alias{init.param.custom}
\title{SVD initialization}
\usage{
init.param.custom(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = poisson(),
  values = list(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
