% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.param.random}
\alias{init.param.random}
\title{Random initialization}
\usage{
init.param.random(Y, X = NULL, Z = NULL, ncomp = 2, family = gaussian())
}
\description{
Initialize the parameters of a GMF model sampling them from an independent
Gaussian distribution (see \code{\link{init.gmf.param}} for more details)
}
\keyword{internal}
