% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.param}
\alias{init.param}
\title{Initialization of the generalized matrix factorization model}
\usage{
init.param(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  method = "svd",
  niter = 0,
  values = list(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
