% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.param.svd}
\alias{init.param.svd}
\title{OLS-SVD initialization}
\usage{
init.param.svd(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = poisson(),
  niter = 0,
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
