% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.param.svd2}
\alias{init.param.svd2}
\title{Modified OLS-SVD initialization}
\usage{
init.param.svd2(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = gaussian(),
  type = c("deviance", "pearson", "working", "link"),
  verbose = FALSE
)
}
\description{
Initialize the parameters of a GMF model fitting a sequence of multivariate
linear regression followed by a residual SVD decomposition. It allows to
recursively refine the initial estimate by repeating the process a pre-specified
number of times. See \code{\link{init.param}} for more details.
}
\keyword{internal}
