% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGMF-class.R
\docType{class}
\name{initgmf-class}
\alias{initgmf-class}
\title{S3 class: initgmf}
\description{
A short description...
}
\section{Slots}{

\describe{
\item{\code{method}}{initialization method to approximatelly minimize the negative penalized log-likelihood}

\item{\code{family}}{a \code{glm} family (see \code{\link{family}} for more details)}

\item{\code{ncomp}}{rank of the latent matrix factorization}

\item{\code{type}}{residual type to be used for initializing the latent scores}

\item{\code{verbose}}{if \code{TRUE}, print the optimization status (default \code{TRUE})}

\item{\code{parallel}}{if \code{TRUE}, allows for parallel computing using the package \code{foreach} (only if \code{method="glm"})}

\item{\code{nthreads}}{number of cores to be used in parallel (only if \code{parallel=TRUE})}

\item{\code{Y}}{matrix of responses (\eqn{n \times m})}

\item{\code{X}}{matrix of row fixed effects (\eqn{n \times p})}

\item{\code{Z}}{matrix of column fixed effects (\eqn{m \times q})}

\item{\code{A}}{matrix of row-specific regression effects (\eqn{n \times q})}

\item{\code{B}}{matrix of column-specific regression effects (\eqn{m \times p})}

\item{\code{U}}{matrix of latent scores (\eqn{n \times d})}

\item{\code{V}}{matrix of factor loadings (\eqn{m \times d})}

\item{\code{phi}}{scalar dispersion parameter}
}}

