% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{ls.svd.init}
\alias{ls.svd.init}
\title{OLS-SVD initialization}
\usage{
ls.svd.init(
  Y,
  X = NULL,
  Z = NULL,
  d = min(dim(Y)),
  family = poisson(),
  niter = 0,
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
