% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.pos.diag}
\alias{make.pos.diag}
\title{Fix sign ambiguity of eigen-vectors}
\usage{
make.pos.diag(U)
}
\description{
Fix sign ambiguity of eigen-vectors by making U positive diagonal
}
\keyword{internal}
