% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{matrix.deviance}
\alias{matrix.deviance}
\title{Model deviance of a GMF model}
\usage{
matrix.deviance(mu, y, family = gaussian())
}
\description{
Compute the overall deviance averaging the contributions of all data
}
\keyword{internal}
