% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{matrix.penalty}
\alias{matrix.penalty}
\title{Frobenius penalty for the parameters of a GMF model}
\usage{
matrix.penalty(U, penalty)
}
\description{
Compute the Frobenius penalty for all the parameters in the model
}
\keyword{internal}
