% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize the matrices U and V with respect to X and Z}
\usage{
orthogonalize(X, Z, B, A, U, V)
}
\description{
Orthogonalize U and V with respect to X and Z, respectively, sequentially applying
multivariate least squares and residual SVD.
}
\keyword{internal}
