% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orthogonalize.uv}
\alias{orthogonalize.uv}
\alias{orthogonalize.svd}
\alias{orthogonalize.qr}
\alias{orthogonalize.std}
\title{Normalize the matrices U and V}
\usage{
orthogonalize.uv(
  U,
  V,
  method = c("SVD", "QR", "ZCA", "ZCA-cor", "PCA", "PCA-cor", "Cholesky")
)

orthogonalize.svd(U, V)

orthogonalize.qr(U, V)

orthogonalize.std(U, V, method)
}
\description{
Rotate U and V using either QR or SVD decompositions.
}
\details{
Orthogonalization is implemented using the following methods:
\itemize{
  \item \code{method = "SVD"}: orthogonal \eqn{U} and scaled orthogonal \eqn{V} based on SVD decomposition;
  \item \code{method = "QR"}: orthogonal \eqn{U} and lower triangular \eqn{V} based on QR decomposition;
  \item \code{method = "ZCA"}: standardized \eqn{U} and lower triangular \eqn{V} based on ZCA whitening and QR decomposition;
  \item \code{method = "ZCA-cor"}: uncorrelated \eqn{U} and lower triangular \eqn{V} based on ZCA whitening and QR decomposition;
  \item \code{method = "PCA"}: standardized \eqn{U} and lower triangular \eqn{V} based on PCA whitening and QR decomposition;
  \item \code{method = "PCA-cor"}: uncorrelated \eqn{U} and lower triangular \eqn{V} based on PCA whitening and QR decomposition;
  \item \code{method = "Cholesky"}: standardized \eqn{U} and lower triangular \eqn{V} based on Cholesky whitening and QR decomposition.
}
}
\keyword{internal}
