% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{plot.sgdgmf}
\alias{plot.sgdgmf}
\title{Plot diagnostics for a GMF model}
\usage{
\method{plot}{sgdgmf}(
  object,
  type = c("1", "2", "3", "4", "5", "idx", "fit", "hist", "qq", "ecdf"),
  resid = c("deviance", "pearson", "working", "response", "link"),
  subsample = FALSE,
  sample.size = 500,
  partial = FALSE,
  normalize = FALSE,
  fillna = FALSE,
  bycol = FALSE
)
}
\arguments{
\item{object}{...}

\item{type}{...}

\item{resid}{...}

\item{subsample}{...}

\item{sample.size}{...}

\item{partial}{...}

\item{normalize}{...}

\item{fillna}{...}

\item{bycol}{...}
}
\description{
A short description...
}
