% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{pointwise.deviance}
\alias{pointwise.deviance}
\title{Pointwise deviance of a GMF model}
\usage{
pointwise.deviance(mu, y, family = gaussian())
}
\description{
Compute the pointwise deviance for all the observations in the sample
}
\keyword{internal}
