% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{predict}
\alias{predict}
\alias{predict.sgdgmf}
\title{Predict method for GMF fits}
\usage{
predict(object, newdata = NULL, type = c("link", "response", "terms"))

\method{predict}{sgdgmf}(object, newdata = NULL, type = c("link", "response", "terms"))
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{newdata}{optionally, a list containing new values for \code{X} and \code{Z}}

\item{type}{the type of prediction which should be returned}
}
\value{
\code{NULL}
}
\description{
...
}
