% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{procrustes}
\alias{procrustes}
\title{Procrustes rotation of two configurations}
\usage{
procrustes(X, Y, scale = TRUE, symmetric = FALSE)
}
\arguments{
\item{X}{target matrix}

\item{Y}{matrix to be rotated}

\item{scale}{allow scaling of axes of Y}

\item{symmetric}{if \code{TRUE}, use symmetric Procrustes statistic}
}
\description{
Rotates a configuration to maximum similarity with another configuration
}
\keyword{internal}
