% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{residuals}
\alias{residuals}
\alias{residuals.sgdgmf}
\title{Extract the residuals of a GMF model}
\usage{
residuals(object, type = c("deviance", "pearson", "working", "response"))

\method{residuals}{sgdgmf}(object, type = c("deviance", "pearson", "working", "response"))
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{type}{the type of residuals which should be returned}
}
\value{
\code{NULL}
}
\description{
...
}
