% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{screeplot.sgdgmf}
\alias{screeplot.sgdgmf}
\title{Screeplot for the residuals of a GMF model}
\usage{
\method{screeplot}{sgdgmf}(
  object,
  ncomp = 20,
  partial = FALSE,
  normalize = FALSE,
  cumulative = FALSE,
  proportion = FALSE
)
}
\arguments{
\item{object}{...}

\item{ncomp}{...}

\item{partial}{...}

\item{normalize}{...}

\item{cumulative}{...}

\item{proportion}{...}
}
\description{
A short description...
}
