% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.control}
\alias{set.control}
\title{Check and set the control parameters for the selecte optimization algorithm}
\usage{
set.control(method, control)
}
\description{
Check if the input control parameters are allowed and set them to default
values if they are not. Returns a list of well-defined control parameters.
}
\keyword{internal}
