% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.cvopt}
\alias{set.cvopt}
\title{Check and set the cross-validation parameters}
\usage{
set.cvopt(cvopt)
}
\description{
Check if the input cross-validation parameters are allowed and set them to default
values if they are not. Returns a list of well-defined cross-validation parameters.
}
\keyword{internal}
