% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.family}
\alias{set.family}
\title{Check and set the model family}
\usage{
set.family(family)
}
\description{
Check if the model family is well-defined and return it eventualy with a
different family name for compatibility with \code{C++}
}
\keyword{internal}
