% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.init}
\alias{set.init}
\title{Check and set the initialization parameters}
\usage{
set.init(init)
}
\description{
Check if the input initialization parameters are allowed and set them to default
values if they are not. Returns a list of well-defined initialization parameters.
}
\keyword{internal}
