% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.jitter}
\alias{set.jitter}
\title{Set the jittering function for mapping the data (DEPRECATED!)}
\usage{
set.jitter(family)
}
\description{
Return a function f which maps the original data y into the perturbed data
z = f(y) such that z is a real number having the same scale of eta = linkinv(mu).
This is useful for initialization purposes.
}
\keyword{internal}
