% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.mat.offset}
\alias{set.mat.offset}
\title{Check and set the offset matrix}
\usage{
set.mat.offset(O, n, m)
}
\description{
Check if the input offset matrix is well-defined and return the same
matrix without attributes such as row and column names.
}
\keyword{internal}
