% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.mat.weights}
\alias{set.mat.weights}
\title{Check and set the weighting matrix}
\usage{
set.mat.weights(W, n, m)
}
\description{
Check if the input weighting matrix is well-defined and return the same
matrix without attributes such as row and column names.
}
\keyword{internal}
