% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.mat.X}
\alias{set.mat.X}
\title{Check and set the covariate matrix X}
\usage{
set.mat.X(X, n, mat = "X")
}
\description{
Check if the input covariate matrix X is well-defined and return the same
matrix without attributes such as row and column names.
}
\keyword{internal}
