% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.mat.Y}
\alias{set.mat.Y}
\title{Check and set the response matrix Y}
\usage{
set.mat.Y(Y)
}
\description{
Check if the input response matrix is well-defined and return the same
matrix without attributes such as row and column names.
}
\keyword{internal}
