% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdgmf-class.R
\docType{class}
\name{sgdgmf-class}
\alias{sgdgmf-class}
\title{S3 class: sgdgmf}
\description{
A short description...
}
\section{Slots}{

\describe{
\item{\code{method}}{estimation method to minimize the negative penalized log-likelihood}

\item{\code{family}}{a \code{glm} family (see \code{\link{family}} for more details)}

\item{\code{ncomp}}{rank of the latent matrix factorization}

\item{\code{control.init}}{list of control parameters for the initialization (see \code{\link{set.control.init}} for more details)}

\item{\code{control.alg}}{list of control parameters for the optimization (see \code{\link{set.control.alg}} for more details)}

\item{\code{control.cv}}{list of control parameters for the cross-validation (see \code{\link{set.control.cv}} for more details)}

\item{\code{Y}}{matrix of responses (\eqn{n \times m})}

\item{\code{X}}{matrix of row fixed effects (\eqn{n \times p})}

\item{\code{Z}}{matrix of column fixed effects (\eqn{m \times q})}

\item{\code{A}}{matrix of row-specific regression effects (\eqn{n \times q})}

\item{\code{B}}{matrix of column-specific regression effects (\eqn{m \times p})}

\item{\code{U}}{matrix of latent scores (\eqn{n \times d})}

\item{\code{V}}{matrix of factor loadings (\eqn{m \times d})}

\item{\code{eta}}{matrix of linear predictor (\eqn{n \times m})}

\item{\code{mu}}{matrix of fitted means (\eqn{n \times m})}

\item{\code{var}}{matrix of fitted variances (\eqn{n \times m})}

\item{\code{phi}}{scalar dispersion parameter}

\item{\code{penalty}}{penalty  obtained at the end of the estimation}

\item{\code{deviance}}{deviance obtained at the end of the estimation}

\item{\code{objective}}{objective function at the end of the estimation}

\item{\code{aic}}{Akaike information criterion of the estimated model}

\item{\code{bic}}{Bayesian information criterion of the estimated model}

\item{\code{cbic}}{Corrected Bayesian information criterion of the estimated model}

\item{\code{exe.time}}{Final execution time in seconds}

\item{\code{trace}}{data frame collecting the optimization history}

\item{\code{summary.cv}}{data frame collecting the cross-validation history}
}}

