% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-package.R
\docType{package}
\name{sgdGMF-package}
\alias{sgdGMF}
\alias{sgdGMF-package}
\title{sgdGMF: Estimation of generalized matrix factorization models via stochastic gradient descent}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
