% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdgmf.R
\name{sgdgmf}
\alias{sgdgmf}
\title{Factorize a matrix of non-Gaussian observations}
\usage{
sgdgmf(
  Y,
  X = NULL,
  Z = NULL,
  family = poisson(),
  ncomp = 2,
  method = "newton",
  intercept = c(FALSE, FALSE),
  dispersion = FALSE,
  penalty = list(),
  init = list(),
  control = list()
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column fixed effects (\eqn{q \times m})}

\item{family}{a family as in the \code{\link{glm}} interface}

\item{ncomp}{number of random effects to estimate (default 2)}

\item{intercept}{should the model include a row/col intercept? (default FALSE)}

\item{dispersion}{should the method estimate column-specific disperssion parameters? (default FALSE)}

\item{penalty}{list of penalty parameters corresponding to u, v, a and b}

\item{init}{list of initialization options}

\item{control}{list of optimization options}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
