% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{SIC}
\alias{SIC}
\title{Schwarz information criterion}
\usage{
SIC(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{optional parameters}
}
\description{
Generic function calculating the Schwarz information criterion for one or several
fitted model objects for which a log-likelihood, or negative deviance, value can
be obtained, according to the formula \eqn{\text{SIC} = - 2 \,\text{log-likelihood} + p \log(n)/n},
where \eqn{p} represents the number effective degrees of freedom of the fitted model,
i.e. the effective number of parameters, and \eqn{n} is the number of observations.
}
