% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{SIC.sgdgmf}
\alias{SIC.sgdgmf}
\title{Compute the SIC of a GMF model}
\usage{
\method{SIC}{sgdgmf}(object)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\description{
Compute the Bayesian information criterion (SIC) of an estimated GMF object
}
