% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{simulate}
\alias{simulate}
\title{Generic function to simulate new data from a statistical model}
\usage{
simulate(object, ...)
}
\description{
Generic function to simulate new data from a statistical model
}
\keyword{internal}
