% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_airwls.R
\name{slice.glm}
\alias{slice.glm}
\title{GLM update computed slice-wise}
\usage{
slice.glm(
  X,
  Y,
  slices,
  coefs,
  offset = NULL,
  penalized = 0,
  parallel = 1,
  family = poisson(),
  method = "step",
  stepsize = 1
)
}
\description{
...
}
\keyword{internal}
