% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{storedata}
\alias{storedata}
\title{Store data into an object}
\usage{
storedata(object, ...)
}
\arguments{
\item{object}{an object from which simulate new data}

\item{...}{additional arguments passed to or from other methods}
}
\value{
An object of the same class as the input containing new data
}
\description{
Generic function to store data into an object, typically a statistical model
}
