% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{storedata.sgdgmf}
\alias{storedata.sgdgmf}
\title{Save response and covariate data into an empty sgdGMF object}
\usage{
\method{storedata}{sgdgmf}(object, Y, X = NULL, Z = NULL)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column fixed effects (\eqn{q \times m})}
}
\value{
A \code{sgdgmf} object containing a copy of the data
}
\description{
Save response and covariate data into an empty sgdGMF object
}
\examples{
\donttest{# Load the sgdGMF package
library(sgdGMF)

# Generate data from a Poisson model
data = sim.gmf.data(n = 100, m = 20, ncomp = 5, family = poisson())

# Fit a GMF model without storing a copy of the data
gmf = sgdgmf.fit(data$Y, ncomp = 3, family = poisson(),
                 control.alg = list(savedata = FALSE))

cat("savedata:", gmf$control.alg$savedata, "\n")
cat("Is Y null?", is.null(gmf$Y), "\n")
cat("Is X null?", is.null(gmf$X), "\n")
cat("Is Z null?", is.null(gmf$Z), "\n")
cat("Is eta null?", is.null(gmf$eta), "\n")
cat("Is mu null?", is.null(gmf$mu), "\n")
cat("Is var null?", is.null(gmf$var), "\n")

# Store the data in the GMF object a posteriori
gmf = storedata(gmf, data$Y)

cat("savedata:", gmf$control.alg$savedata, "\n")
cat("Y:", dim(gmf$Y), "\n")
cat("X:", dim(gmf$X), "\n")
cat("Z:", dim(gmf$Z), "\n")
cat("eta:", dim(gmf$eta), "\n")
cat("mu:", dim(gmf$mu), "\n")
cat("var:", dim(gmf$var), "\n")
}
}
