% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.sgdgmf}
\title{Summarizing a GMF model}
\usage{
summary(object)

\method{summary}{sgdgmf}(object)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\value{
\code{NULL}
}
\description{
...
}
