% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_newton.R
\name{update.newton}
\alias{update.newton}
\title{Newton update for the parameters of a GMF model}
\usage{
\method{update}{newton}(U, V, penalty, ddiff, ddratio, stepsize, damping)
}
\description{
Compute the Newton updates for U given V
}
\keyword{internal}
