% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vglmfit.R
\name{vglm.fit.coef}
\alias{vglm.fit.coef}
\title{Estimate the coefficients of a vector generalized linear model}
\usage{
vglm.fit.coef(
  Y,
  X,
  family = gaussian(),
  offset = NULL,
  parallel = FALSE,
  nthreads = 1,
  clust = NULL
)
}
\description{
...
}
\keyword{internal}
