% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whitening.matrix}
\alias{whitening.matrix}
\title{Compute the whitening matrix from a given covariance matrix}
\usage{
whitening.matrix(
  sigma,
  method = c("ZCA", "ZCA-cor", "PCA", "PCA-cor", "Cholesky")
)
}
\arguments{
\item{sigma}{covariance matrix.}

\item{method}{determines the type of whitening transformation.}
}
\description{
Compute the whitening matrix from a given covariance matrix
}
\keyword{internal}
