% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{biplot.sgdgmf}
\alias{biplot.sgdgmf}
\title{Biplot of a GMF model}
\usage{
\method{biplot}{sgdgmf}(
  object,
  choices = 1:2,
  normalize = FALSE,
  labels = NULL,
  palette = c("viridis", "inferno")
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{choices}{a length 2 vector specifying the components to plot}

\item{normalize}{if \code{TRUE}, orthogonalizes the scores using SVD}

\item{labels}{a vector of labels which should be plotted}

\item{palette}{the name of a color-palette which should be used}
}
\description{
Plot the observations on a two-dimensional projection determined by the
estimated score matrix
}
