% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{coefficients.sgdgmf}
\alias{coefficients.sgdgmf}
\alias{coef.sgdgmf}
\title{Extract the coefficient of a GMF model}
\usage{
\method{coefficients}{sgdgmf}(object, type = c("all", "colreg", "rowreg", "scores", "loadings"))

\method{coef}{sgdgmf}(object, type = c("all", "colreg", "rowreg", "scores", "loadings"))
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{type}{the type of coefficients which should be returned}
}
\description{
Return the estimated coefficients of a GMF model, i.e., the row- and column-specific
regression effects, the latent scores and loadings.
}
