% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{eigenval}
\alias{eigenval}
\title{Generic function to compute the eigenvalues of a matrix}
\usage{
eigenval(object, ...)
}
\description{
Generic function to compute the eigenvalues of a matrix
}
\keyword{internal}
