% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{eigenval.sgdgmf}
\alias{eigenval.sgdgmf}
\title{Spectrum method for GMF models}
\usage{
eigenval.sgdgmf(
  object,
  ncomp = object$ncomp,
  type = c("deviance", "pearson", "working", "link"),
  normalize = FALSE
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{ncomp}{number of eigenvalues to compute}

\item{type}{the type of residual which should be returned}

\item{normalize}{if \code{TRUE}, standardize the residuals column-by-column}
}
\description{
Compute the latent spectrum of a GMF model evaluated on the GLM residual scale.
}
\details{
Let \eqn{g(\mu) = \eta = X B^\top + \Gamma Z^\top} be the linear predictor of a
GMF model where the latent factorization \eqn{U V^\top} is not included into the
model specification. Let \eqn{R} be the correspondent partial residual matrix.
Either deviance, \eqn{r_{ij} = \textrm{sign}(y_{ij} - \mu_{ij}) \sqrt{D(y_{ij}, \mu_{ij})}},
or Pearson, \eqn{r_{ij} = (y_{ij} - \mu_{ij}) / \sqrt{\nu(\mu_{ij})}}, residuals
can be considered. Finally, we define \eqn{\Sigma} as the empirical variance-covariance
matrix of \eqn{R}, being \eqn{\sigma_{ij} = \textrm{Cov}(r_{:i}, r_{:j})}. Then, we define
the latent spectrum of the model as the collection of eigenvalues of \eqn{\Sigma}.
Notice that, in case of Gaussian data, the latent spectrum corresponds to the principal
component analysis on the regression residuals, whose eigenvalues can be used to
infer how much signal can be explained by each principal component. Similarly,
we can use the latent spectrum in non-Gaussian data settings to infer the correct
number of principal components to include into the GMF model.
}
\keyword{internal}
