% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{heatmap.sgdgmf}
\alias{heatmap.sgdgmf}
\title{Heatmap of a GMF model}
\usage{
\method{heatmap}{sgdgmf}(
  object,
  type = c("data", "response", "link", "scores", "loadings", "deviance", "pearson",
    "working"),
  resid = FALSE,
  symmetric = FALSE,
  transpose = FALSE,
  limits = NULL,
  palette = NULL
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{type}{the type of data/predictions/residuals which should be returned}

\item{resid}{if \code{TRUE}, plots the residual values}

\item{symmetric}{if \code{TRUE}, symmetrizes the color limits}

\item{transpose}{if \code{TRUE}, transposes the matrix before plotting it}

\item{limits}{the color limits which should be used}

\item{palette}{the color-palette which should be used}
}
\description{
Plots a heatmap of either the data, the fitted values, or the residual values
of a GMF model allowing for different types of transformations and normalizations.
Moreover, it also permits to plot the latent score and loading matrices.
}
