% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init.param.custom}
\alias{init.param.custom}
\title{SVD initialization}
\usage{
init.param.custom(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = poisson(),
  values = list(),
  verbose = FALSE
)
}
\description{
Initialize the parameters of a GMF model using custom values provided by the user
and estimating the unspecified parameters using the same procedure described in
\code{\link{init.param.svd}}. See \code{\link{init.param}} for more details.
}
\keyword{internal}
