% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init.param.glm}
\alias{init.param.glm}
\title{GLM-SVD initialization}
\usage{
init.param.glm(
  Y,
  X = NULL,
  Z = NULL,
  ncomp = 2,
  family = poisson(),
  verbose = FALSE
)
}
\description{
...
}
\keyword{internal}
