% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGMF-class.R
\docType{class}
\name{initgmf-class}
\alias{initgmf-class}
\title{S3 class: initgmf}
\description{
A short description...
}
\section{Slots}{

\describe{
\item{\code{method}}{initialization method}

\item{\code{family}}{model family}

\item{\code{ncomp}}{number of latent components}

\item{\code{type}}{residual type}

\item{\code{verbose}}{if \code{TRUE}, ...}

\item{\code{parallel}}{if \code{TRUE}, ...}

\item{\code{nthreads}}{...}

\item{\code{A}}{matrix of row-specific regression effects (\eqn{n \times q})}

\item{\code{B}}{matrix of column-specific regression effects (\eqn{m \times p})}

\item{\code{U}}{matrix of latent scores (\eqn{n \times d})}

\item{\code{V}}{matrix of factor loadings (\eqn{m \times d})}

\item{\code{phi}}{scalar dispersion parameter}
}}

