% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.uv}
\alias{normalize.uv}
\title{Normalize the matrices U and V}
\usage{
normalize.uv(U, V, method = c("svd", "qr"))
}
\description{
Rotate U and V in such a way that the transformed matrices
are such that U is orthogonal and V is lower triangular
}
\keyword{internal}
