% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalize the matrices U and V with respect to X and Z}
\usage{
orthogonalize(
  X,
  Z,
  B,
  A,
  U,
  V,
  method = c("QR", "SVD", "ZCA", "ZCA-cor", "PCA", "PCA-cor", "Cholesky")
)
}
\description{
Orthogonalize \code{[A, U]} and \code{V} with respect to \code{X} and \code{Z},
respectively, sequentially applying multivariate least squares and residual
whitening on U. The result must satisfy the following contraints:
\eqn{X^\top A = 0}, \eqn{X^\top U = 0}, \eqn{Z^\top V = 0}, \eqn{U^\top U = 0}.
}
\keyword{internal}
