% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{plot.sgdgmf}
\alias{plot.sgdgmf}
\title{Plot diagnostics for a GMF model}
\usage{
\method{plot}{sgdgmf}(
  object,
  type = c("1", "2", "3", "4", "5", "6", "idx", "fit", "std", "hist", "qq", "ecdf"),
  resid = c("deviance", "pearson", "working", "response", "link"),
  subsample = FALSE,
  sample.size = 500,
  partial = FALSE,
  normalize = FALSE,
  fillna = FALSE,
  bycol = FALSE
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{type}{the type of plot which should be returned}

\item{resid}{the type of residuals which should be used}

\item{subsample}{if \code{TRUE}, computes the residuals over o small fraction of the data}

\item{sample.size}{the dimension of the sub-sample which should be used}

\item{partial}{if \code{TRUE}, computes the partial residuals}

\item{normalize}{if \code{TRUE}, standardizes the residuals column-by-column}

\item{fillna}{if \code{TRUE}, fills the \code{NA} values with \code{0}}

\item{bycol}{if \code{TRUE}, uses a column-specific palette}
}
\description{
Plots (one of) six diagnostics to graphically analyze the marginal and conditional
distribution of the residuals of a GMF model. Currently, the following plots are
available: residuals against observation indices, residuals agains fitted values,
absolute square-root residuals against fitted values, histogram of the residuals,
residual QQ-plot, residual ECDF-plot.
}
