% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sgdgmf}
\alias{predict.sgdgmf}
\title{Predict method for GMF fits}
\usage{
\method{predict}{sgdgmf}(
  object,
  newdata = NULL,
  idx = NULL,
  type = c("link", "response", "terms"),
  ...
)
}
\description{
...
}
