% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdgmf-class.R
\name{print.sgdgmf}
\alias{print.sgdgmf}
\title{Print the fundamental characteristics of a GMF}
\usage{
\method{print}{sgdgmf}(object)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}
}
\description{
...
}
