% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{refit.sgdgmf}
\alias{refit.sgdgmf}
\title{Refine the final estimate of a GMF model}
\usage{
\method{refit}{sgdgmf}(
  object,
  ...,
  normalize = TRUE,
  verbose = FALSE,
  parallel = FALSE,
  nthreads = 1
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{...}{further arguments passed to or from other methods}

\item{normalize}{if \code{TRUE}, normalize \code{U} and \code{V} to uncorrelated Gaussian \code{U} and upper triangular \code{V} with positive diagonal}

\item{verbose}{if \code{TRUE}, print the optimization status}

\item{parallel}{if \code{TRUE}, use parallel computing using the \code{foreach} package}

\item{nthreads}{number of cores to be used in the \code{"glm"} method}
}
\description{
Refine the estimated latent scores of a GMF model via IRWLS
}
\examples{
library(sgdGMF)

# Set RUN = TRUE to run the example
RUN = FALSE
if (RUN) {
  # Generate data from a Poisson model
  data = sim.gmf.data(n = 100, m = 20, ncomp = 5, family = poisson())

  # Fit a GMF model using SGD
  gmf_old = sgdgmf.fit(data$Y, ncomp = 3, family = poisson(), method = "sgd")

  # Refine the score matrix estimate
  gmf_new = refit(gmf_old)

  # Get the fitted values in the link and response scales
  mu_hat_old = fitted(gmf_old, type = "response")
  mu_hat_new = fitted(gmf_new, type = "response")

  # Compare the results
  par(mfrow = c(2,2), mar = c(1,1,3,1))
  image(data$Y, axes = FALSE, main = expression(Y))
  image(data$mu, axes = FALSE, main = expression(mu))
  image(mu_hat_old, axes = FALSE, main = expression(hat(mu)[old]))
  image(mu_hat_new, axes = FALSE, main = expression(hat(mu)[new]))
}

}
