% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{refit.sgdgmf}
\alias{refit.sgdgmf}
\title{Refine the final estimate of a GMF model}
\usage{
\method{refit}{sgdgmf}(
  object,
  normalize = FALSE,
  verbose = FALSE,
  parallel = FALSE,
  nthreads = 1,
  clust = NULL
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{verbose}{...}

\item{parallel}{...}

\item{nthreads}{...}
}
\description{
Refine the estimated latent scores of a GMF model via IRWLS
}
