% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGMF-class.R
\name{residuals.initgmf}
\alias{residuals.initgmf}
\alias{resid.initgmf}
\title{Extract the residuals of an initialized GMF model}
\usage{
\method{residuals}{initgmf}(
  object,
  type = c("deviance", "pearson", "working", "response", "link"),
  partial = FALSE,
  normalize = FALSE,
  fillna = FALSE,
  spectrum = FALSE,
  ncomp = 50
)

\method{resid}{initgmf}(
  object,
  type = c("deviance", "pearson", "working", "response", "link"),
  partial = FALSE,
  normalize = FALSE,
  fillna = FALSE,
  spectrum = FALSE,
  ncomp = 50
)
}
\arguments{
\item{object}{an object of class \code{initgmf}}

\item{type}{the type of residuals which should be returned}

\item{partial}{if \code{TRUE}, computes the residuals excluding the matrix factorization from the linear predictor}

\item{normalize}{if \code{TRUE}, standardize the residuals column-by-column}

\item{fillna}{if \code{TRUE}, fills \code{NA} values column-by-column}

\item{spectrum}{if \code{TRUE}, returns the eigenvalues of the residual covariance matrix}

\item{ncomp}{number of eigenvalues to be calculated (only if \code{spectrum=TRUE})}
}
\description{
Extract the residuals of an initialized GMF model and, if required, compute
the eigenvalues of the residuals covariance/correlation matrix.
Moreover, if required, return the partial residual of the model obtained by
excluding the matrix decomposition from the linear predictor.
}
\seealso{
\code{\link{residuals.sgdgmf}} and \code{\link{resid.sgdgmf}} for more details on the residual computation.
}
