% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-class.R
\name{screeplot.sgdgmf}
\alias{screeplot.sgdgmf}
\title{Screeplot for the residuals of a GMF model}
\usage{
\method{screeplot}{sgdgmf}(
  object,
  ncomp = 20,
  type = c("deviance", "pearson", "working", "response", "link"),
  partial = FALSE,
  normalize = FALSE,
  cumulative = FALSE,
  proportion = FALSE
)
}
\arguments{
\item{object}{an object of class \code{sgdgmf}}

\item{ncomp}{number of components to be plotted}

\item{type}{the type of residuals which should be used}

\item{partial}{if \code{TRUE}, plots the eigenvalues of the partial residuals}

\item{normalize}{if \code{TRUE}, plots the eigenvalues of the standardized residuals}

\item{cumulative}{if \code{TRUE}, plots the cumulative sum of the eigenvalues}

\item{proportion}{if \code{TRUE}, plots the fractions of explained variance}
}
\description{
Plots the variances of the principal components of the residuals against the
number of principal component.
}
