% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.control.cv}
\alias{set.control.cv}
\title{Check and set the cross-validation parameters}
\usage{
set.control.cv(
  criterion = c("dev", "aic", "bic", "cbic"),
  refit = TRUE,
  nfolds = 5,
  proportion = 0.3,
  verbose = FALSE,
  parallel = FALSE,
  nthreads = 1
)
}
\arguments{
\item{nfolds}{number of cross-validation folds}

\item{parallel}{if \code{TRUE}, allows for parallel computing}

\item{nthreads}{number of cores to use in parallel (only if \code{parallel=TRUE})}
}
\description{
Check if the input cross-validation parameters are allowed and set them to default
values if they are not. Returns a list of well-defined cross-validation parameters.
}
