% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.control}
\alias{set.control}
\title{Check and set the control parameters for the select optimization algorithm}
\usage{
set.control(method, control)
}
\arguments{
\item{method}{optimization method: \code{"airwls"}, \code{"newton"}, \code{"msgd"}, \code{"csgd"}, \code{"rsgd"}, \code{"bsgd"}}

\item{control}{list of algorithm-specific control parameters}
}
\description{
Check if the input control parameters are allowed and set them to default
values if they are not. Returns a list of well-defined control parameters.
}
\details{
It is not necessary to provide a complete list of control parameters, one can
just specify a list containing the parameters he/she needs to change from the
default values. Wrongly specified parameters are ignored or set to default values.
For a detailed description of all the algorithm-specific control parameters,
please refer to
\code{\link{set.control.airwls}} (\code{method="airwls"}),
\code{\link{set.control.newton}} (\code{method="newton"}),
\code{\link{set.control.csgd}} (\code{method="csgd"}),
\code{\link{set.control.bsgd}} (\code{method="bsgd"}).
}
\keyword{internal}
