% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.control.rsgd}
\alias{set.control.rsgd}
\title{Check and set the control parameters for the rowwise-SGD algorithm}
\usage{
set.control.rsgd(control)
}
\description{
Check if the input control parameters are allowed and set them to default
values if they are not. Returns a list of well-defined control parameters.
}
\keyword{internal}
