% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.family}
\alias{set.family}
\title{Check and set the model family}
\usage{
set.family(family)
}
\arguments{
\item{family}{a \code{glm} family (see \code{\link{family}} for more details)}
}
\description{
Check if the model family is allowed and return it eventually with a
different family name for compatibility with the \code{C++} implementation
}
\keyword{internal}
