% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set.penalty}
\alias{set.penalty}
\title{Check and set the penalty parameters}
\usage{
set.penalty(penalty)
}
\description{
Check if the input penalty parameters are allowed and set them to default
values if they are not. Returns a list of well-defined penalty parameters.
}
\keyword{internal}
