\name{sgdGMF-package}
\alias{sgdGMF-package}
\alias{sgdGMF}
\docType{package}
\title{
\packageTitle{sgdGMF}
}
\description{
\packageDescription{sgdGMF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sgdGMF}
\packageIndices{sgdGMF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sgdGMF}

Maintainer: \packageMaintainer{sgdGMF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
