% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdGMF-package.R
\docType{package}
\name{sgdGMF-package}
\alias{sgdGMF}
\alias{sgdGMF-package}
\title{sgdGMF: Estimation of generalized matrix factorization models via stochastic gradient descent}
\description{
The sgdGMF package provides an efficient infrastructure to estimate high-dimensional generalized matrix factorization (GMF) models using penalized maximum likelihood under a dispersion exponential family specification. Either deterministic and stochastic methods are implemented for the numerical maximization. In particular, the package implements an original stochastic gradient descent (SGD) algorithm with a block-wise minibatch strategy to speed up the computations and an efficient adaptive learning rate schedule to stabilize the convergence. Other methods considered for the optimization are the alternated iterative re-weighted least squares (AIRWLS) and the quasi-Newton method with diagonal approximation of the Fisher information matrix.
}
\keyword{internal}
