% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{sgdgmf.cv}
\alias{sgdgmf.cv}
\title{Model selection via cross-validation for generalized matrix factorization models}
\usage{
sgdgmf.cv(
  Y,
  X = NULL,
  Z = NULL,
  family = poisson(),
  ncomp = 2,
  method = c("airwls", "newton", "msgd", "csgd", "rsgd", "csgd"),
  penalty = list(),
  control.init = list(),
  control.alg = list(),
  control.cv = list()
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column fixed effects (\eqn{q \times m})}

\item{family}{a family as in the \code{\link{glm}} interface}

\item{ncomp}{number of random effects to estimate (default 2)}

\item{method}{optimization method: \code{"airwls"} (default),
\code{"newton"}, \code{"msgd"}, \code{"csgd"}, \code{"rsgd"}, \code{"bsgd"}}

\item{penalty}{list of penalty parameters corresponding to u, v, a and b}

\item{nfolds}{number of cross-validation folds (default 5)}

\item{parallel}{if \code{TRUE}, use parallel \code{foreach} to fit the models over different folds}

\item{init}{list of initialization options}

\item{control}{list of optimization options}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
