% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{sgdgmf.cv}
\alias{sgdgmf.cv}
\title{Model selection via cross-validation for generalized matrix factorization models}
\usage{
sgdgmf.cv(
  Y,
  X = NULL,
  Z = NULL,
  family = poisson(),
  ncomps = seq(from = 1, to = 10, by = 1),
  method = c("airwls", "newton", "msgd", "csgd", "rsgd", "csgd"),
  penalty = list(),
  control.init = list(),
  control.alg = list(),
  control.cv = list()
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column fixed effects (\eqn{q \times m})}

\item{family}{a \code{glm} family (see \code{\link{family}} for more details)}

\item{ncomps}{rank of the latent matrix factorization (default 2)}

\item{method}{estimation method to minimize the negative penalized log-likelihood}

\item{penalty}{list of penalty parameters (see \code{\link{set.penalty}} for more details)}

\item{control.init}{list of control parameters for the initialization (see \code{\link{set.control.init}} for more details)}

\item{control.alg}{list of control parameters for the optimization (see \code{\link{set.control.alg}} for more details)}

\item{control.cv}{list of control parameters for the cross-validation (see \code{\link{set.control.cv}} for more details)}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{
...

}
\references{
...
}
