% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{sgdgmf.cv.step}
\alias{sgdgmf.cv.step}
\title{Single step of cross-validation for generalized matrix factorization models}
\usage{
sgdgmf.cv.step(
  train,
  test,
  X,
  Z,
  family,
  ncomp,
  maxcomp,
  fold,
  nfolds,
  method,
  penalty,
  control.init,
  control.alg,
  control.cv
)
}
\value{
Returns a \code{data.frame}  containing the current number of latent factors
in the model (\code{ncomp}), the fold identifier (\code{fold}), the degrees of
freedom, i.e. the number of parameters, of the model (\code{df}), the AIC, BIC
and SIC and deviance (respectively, \code{aic}, \code{bic}, \code{sic}, \code{dev})
calculated on the train and test sets.
}
\description{
Internal function running a single step of cross-validation for generalized
matrix factorization (GMF) models and calculating some goodness-of-fit measures
on the train and test sets.
}
\keyword{internal}
