% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigengap.R
\name{sgdgmf.rank}
\alias{sgdgmf.rank}
\title{Rank selection via eigenvalue-gap methods}
\usage{
sgdgmf.rank(
  Y,
  X = NULL,
  Z = NULL,
  maxcomp = ncol(Y),
  family = gaussian(),
  weights = NULL,
  offset = NULL,
  method = c("onatski", "act", "oht"),
  type.reg = c("ols", "glm"),
  type.res = c("deviance", "pearson", "working", "link"),
  normalize = FALSE,
  maxiter = 10,
  parallel = FALSE,
  nthreads = 1,
  return.eta = FALSE,
  return.mu = FALSE,
  return.res = FALSE,
  return.cov = FALSE
)
}
\arguments{
\item{Y}{matrix of responses (\eqn{n \times m})}

\item{X}{matrix of row-specific fixed effects (\eqn{n \times p})}

\item{Z}{matrix of column-specific fixed effects (\eqn{q \times m})}

\item{maxcomp}{maximum number of eigenvalues to compute}

\item{family}{a family as in the \code{\link{glm}} interface (default \code{gaussian()})}

\item{weights}{matrix of optional weights (\eqn{n \times m})}

\item{offset}{matrix of optional offsets (\eqn{n \times m})}

\item{method}{rank selection method}

\item{type.reg}{regression method to be used to profile out the covariate effects}

\item{type.res}{residual type to be decomposed}

\item{normalize}{if \code{TRUE}, standardize column-by-column the residual matrix}

\item{maxiter}{maximum number of iterations}

\item{parallel}{if \code{TRUE}, allows for parallel computing using \code{foreach}}

\item{nthreads}{number of cores to be used in parallel (only if \code{parallel=TRUE})}

\item{return.res}{if \code{TRUE}, return the residual matrix}

\item{return.cov}{if \code{TRUE}, return the covariance matrix of the residuals}

\item{return.fit}{if \code{TRUE}, return the fitted value martix}
}
\description{
Select the number of significant principal components of a GMF model via
exploitation of eigenvalue-gap methods
}
\examples{
library(sgdGMF)

# Set the data dimensions
n = 100; m = 20; d = 5

# Set the time range, phases, frequences and amplitudes of the underlying signals
time = seq(from = 0, to = 1, length = n)
phase = sort(2*runif(d), decreasing = FALSE)
freq = sort(2*runif(d), decreasing = FALSE)
amp = sort(runif(d), decreasing = TRUE)

Time = tcrossprod(time, rep(1, d))
Phase = tcrossprod(rep(1, n), phase)
Freq = tcrossprod(rep(1, n), freq)
Amp = diag(amp)

# Combine the latent signals using independent Gaussian coefficients
U = sin(2 * pi * Freq * (Time + Phase)) \%*\% Amp
V = matrix(rnorm(m*d), nrow = m, ncol = d)
eta = tcrossprod(U, V)

# Generate data using Poisson, Binomial and Gamma models
mu_pois = exp(eta)
mu_bin = plogis(2*eta)
mu_gam = exp(eta)

Y_pois = matrix(rpois(n*m, lambda = mu_pois), nrow = n, ncol = m)
Y_bin = matrix(rbinom(n*m, size = 1, prob = mu_bin), nrow = n, ncol = m)
Y_gam = matrix(rgamma(n*m, shape = 2, scale = mu_gam), nrow = n, ncol = m)

# Initialize the GMF parameters assuming 3 latent factors
ncomp_pois = sgdgmf.rank(Y_pois, family = poisson(), normalize = TRUE)
ncomp_bin = sgdgmf.rank(Y_bin, family = binomial(), normalize = TRUE)
ncomp_gam = sgdgmf.rank(Y_gam, family = Gamma(link = "log"), normalize = TRUE)

# Get the selected number of components
print(paste("Poisson:", ncomp_pois$ncomp))
print(paste("Binomial:", ncomp_bin$ncomp))
print(paste("Gamma:", ncomp_gam$ncomp))

# Plot the screeplot used for the component determination
par(mfrow = c(3,1))
barplot(ncomp_pois$lambdas, main = "Poisson screeplot")
barplot(ncomp_bin$lambdas, main = "Binomial screeplot")
barplot(ncomp_gam$lambdas, main = "Gamma screeplot")
par(mfrow = c(1,1))
}
\references{
Onatski, A. (2010).
\emph{Determining the number of factors from empirical distribution of eigenvalues.}
Review of Economics and Statistics, 92(4): 1004-1016

Gavish, M., Donoho, D.L. (2014)
\emph{The optimal hard thresholding for singular values is 4/sqrt(3).}
IEEE Transactions on Information Theory, 60(8): 5040--5053

Fan, J., Guo, J. and Zheng, S. (2020).
\emph{Estimating number of factors by adjusted eigenvalues thresholding.}
Journal of the American Statistical Association, 117(538): 852--861

Wang, L. and Carvalho, L. (2023).
\emph{Deviance matrix factorization.}
Electronic Journal of Statistics, 17(2): 3762-3810
}
