% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericfun.R
\name{simulate}
\alias{simulate}
\title{Simulate new data}
\usage{
simulate(object, ...)
}
\arguments{
\item{object}{an object from which simulate new data}

\item{...}{additional arguments passed to or from other methods}
}
\value{
An array containing the simulated data.
}
\description{
Generic function to simulate new data from a statistical model
}
