% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whitening.matrix}
\alias{whitening.matrix}
\alias{whitening.zca}
\alias{whitening.zca.cor}
\alias{whitening.pca}
\alias{whitening.pca.cor}
\alias{whitening.chol}
\title{Compute the whitening matrix from a given covariance matrix}
\usage{
whitening.matrix(
  sigma,
  method = c("ZCA", "ZCA-cor", "PCA", "PCA-cor", "Cholesky")
)

whitening.zca(sigma)

whitening.zca.cor(sigma)

whitening.pca(sigma)

whitening.pca.cor(sigma)

whitening.chol(sigma)
}
\arguments{
\item{sigma}{covariance matrix.}

\item{method}{determines the type of whitening transformation.}
}
\description{
Compute the whitening matrix from a given covariance matrix
}
\details{
This function is an internal re-implementation of the function \code{whiteningMatrix}
in the \code{whitening} package. See the original documentation to get more details.
}
\keyword{internal}
